Resolve more "let" parsing bugs
Find and resolve more parsing bugs with the let function. And, it turns
out that the accidentally "unresolved" address error can potentially show
up at a location a variable/constant will be kept at. Clear that flag
when establishing the location, to prevent it from happening.

Also, zero-pad the numbers when generating output.