Add integertriangle solver.
I was working on trying to present novel ways to display a number
to my students, and I was hoping to show it as the long side of
a triangle. But, this was best limited to integer solutions, so
this was written to quickly check if there was an integer solution.