Comment on broken isPrime
Since it doesn't cause problems with this program, it's not being fixed. But, the note was made about how the isPrime function does not recognize 2 as prime.  We're only concerned about 4 digit primes here, so it's not an issue.  Would be an issue in other areas, if I resuse
this function where 2 would need to be properly recognized.