Here's day 4 done.
This one was a little bit harder, but that was because I had a stupid
test condition backwards. Finding that, for part 1, was hard because
I wrote most of the code in one big braindump.

The second part is almost complete, but I didn't want to write the
string matching code. I just printed all the valid names and then
used grep on the output.