Day 10 part 2 done
This actually shows the solution for both 1 and 2.  I went and
used negative numbers to show actual outputs and not bots. The
existence of a bot #0 and an output #0 caused a big headache.
I solved that by making the outputs store are one less than
their negative number. This means 0=-1 amd 1=-2.  This allowed
me to distinguish between outputs and bots.

A more correct answer would involve having the bot keep track of
the destination type.  And then handle that.  Oh well. It is
solved.