Done with day 21
I figured the easiest way to do the second part would be to
reverse the directions and then rewrite each function so that
it would undo what it normally would have done.  For most of
these, that was the same thing. But, a couple of them needed
to be changed.

That did work.  That seemed easier than checking all possible
combinations or other brute force methods.