I finished this.
I originally submitted the location, not the value, on the website.
Which told me it was wrong. So, I spent an inordinate amount of time
looking for a potential bug in how I was calculating the next value,
and calculating the values by hand to compare to my output.  LOL

Turns out the correct answer was sitting right on the screen while
I did this.  I just needed to put that number in on the website.