Day 07 part a
OK, this was a bit of a cheat. I just parsed the file to remove
everything but the names, putting each word on its own line. Then
I sorted it, and used the uniq command to count repeats, then I
sorted again to find the only word that shows up a single time. That
has to be the bottom node.