This is now done
The second part was harder.  I originally tried to solve it as if
each position would always end up in the same spot.  When it did
not, I went and tried to optimise by creating a command list (to
avoid the cost of text processing).  This was still way too slow
for 1 billion iterations.

The obvious solution (of checking for when we get back to the
start) was the final breakthrough.  After that.  Simple.