Day 23
The second part was fun. I liked forcing myself to work out what
it was actually trying to do.  Once I realized that it was
checking for multiples and incrementing when it found one. It
struck me that it was looking for non-primes. And it was just
looking from one number to the other with a step of 17.

That was a trivial cpp program to write.