Added dc program
This one answers what is the first value of b that gives a lower
number when a = 1.  This code is very obfuscated. It helps that
dc allows a register to be any character. This means we can define
register 9 to hold the value of 2.  Or register " to hold a value