Change playlist to quit early when music=no
The main change here is to put an early exit in playlist so it stops
changing the volume when it shouldn't be running.  There are minor
time changes for when songs play on the tabfile. And there is a temp
change to the lights program so the kids could see a smaller number
run all the way through.