Improve code readability and simplify error checks
I added a specific function to check that the next token is a valid
token, and it will check that it both exists and comes from the
same line of the source file. This permits printing the current line
that has the error at the time we find it.  It also keeps the error
checking logic out if the logic for compiling each instruction.

I also changed the way I labelled errors. TRUE and FALSE were
confusing. A function that failed returned TRUE. It is better for
it to return BAD. But, now "error" made little sense for that
variable name. It was renamed to compile_status.

More constants were consolidated into the same enum, and some constants
were moved out of the INSTSET enum, as they weren't instructions.