Improve synacor stack and coin.s program
In order to increase speed in coin.s, I tried to improve the stack
in the simulator to avoid calling malloc often. I even pushed a dead
value at the start of the program to keep the stack in memory. There
was no noticeable improvement in the runtime of coin.s with that
change. I went back and looked over the code again. I realized that
I didn't need to recalculate the sum every time I added a penny. I
could just increase the calculated sum at the same time I added the
penny. I also took all the memory reads and writes out of that inner
loop.  This cut the runtime by more than half. The fastdivide cut
things down by about 25%, and this is another 50%.  It went from
taking more than 42 seconds to running under 15 seconds.