Optimize coin.s more
Down to 12 seconds for 499 cents. Added another array to keep count
of coins in parallel with the value of the count. This saves the
division step when printing them out as I can print the counts. It
saves us from multiplying when finding the sum as we can just add
the values.  It isn't a huge increase in speed. I also reversed the
look for adding the values up, so we don't have to break the data
lines. It's a cosmetic fix, but that was bothering me.