Shave instructions from pnumber and pstr
We use 0 as termination markers, so don't use eq and then jt, just
reverse the test and use jt with the register and fall through when
it is 0.  This saves us a register to push/pop, an eq compare, and
a jump instruction.  It's minor, but these are two of the more
commonly called functions and every instruction shaved off is
multiplied hundreds of thousands of times when run with a large
number of cents in coin.s