Revise the simulator stack
Instead of allowing a dynamic stack, use a fixed size instead. This
does create a possible crash situation, but I picked a constant that
should be higher than we ever need. This seems to speed things up
dramatically. I wrote all the functions to be inline, so that the
stack access is fast and integrated.  I added some accessor functions
to get certain debug code working.