Compiler change to how strings are supported
We require that we manually null terminate strings. This will
allow finer control when writing the assembly. This helps with
longer messages and saves us function calls.

I also inlined get_add and set_add in the simulator. This speeds
up access to each memory location and cut our time down even
further for the coin.s program.

The krypto program is mostly complete. There is a possibility
that it will need to be cleaned up and made nicer.