Print an error message on bad entry
krypto.solver2 will now print an error message when the number
entered is not between 1 and 99 inclusive. This is more
userfriendly, and it also has the benefit of preventing the string
buffers from overflowing and colliding when someone tries entering
a 3 digit number, when they are sized for two digit numbers as the
maximum length.