remove calls to code used only once
In the krypto solver program, we remove the calls for code that is
used only once. We put them in the right location instead. We also
take strcmp and make it more tightly integrated into the main loop.
This is helpful because it removes code from the part of the
program that is used the most often.