fix the overflow issue in ksolve-final
This fixes the issue in two different ways.  First, it moves the
array information lower in memory.  This keeps the program below
the data for the bad case of "1 2 3 4 5 4".  But, we also have
a check to make sure we don't overflow. If we do, it prints a
message and does not save the current string.