Both parts are done.
This is a little tricky, because you need to execute only those
functions that make sense for your current situation. You load
each function as required.  Then only load the isvalidp for the
part of the day you are doing (a or b).  Then you can call process.

You also can load the testing data or the real data.