Day 3 complete
You can load this entire file and it should execute automatically.
It is important to make sure you open emacs in the same folder where
the data files are stored, or else you need to change the path for
the files to have them load properly.

This is a cleaner implementation than the other two days, but I still
feel like I can improve a lot with how I approach things.