Start working on day 5
We're going to process the input file into a list of seat numbers
in the form ((row# col#) (row# col#) ...) so that we can then go
through and calculate each one to find the max. At this time, I have
got the function for finding the max ID done, but I need to process
the seat descriptions to generate the list.