Finish with Day 5 Part 1
This is probably not optimal, as the seat ID is just the value
of the binary string where B and R are 1s and F and L are 0s.

But, I still am keeping this just in case we really need to have
the numbers different for the next part.